requirements/base.txt
Django==2.1.2
djangorestframework==3.9.0
django-extensions==2.1.3
djangorestframework-camel-case==0.2.0